Fixed testing bugs and Working on setting modules

+++ database/migrations/2025_09_03_110100_create_stock_entries_table.php
@@ -31,7 +31,7 @@ return new class extends Migration
                 ->constrained('users')
                 ->cascadeOnDelete();
 
-            $table->integer('quantity')->default(0);
+            $table->decimal('quantity', 10, 2)->default(0);
             $table->decimal('price', 10, 2)->default(0);
             $table->decimal('value', 12, 2)->default(0);
